+++ async_pokepy/__init__.py
@@ -13,7 +13,7 @@ __author__ = "Lorenzo"
 __docformat__ = "restructuredtext en"
 __license__ = "MIT"
 __copyright__ = "Copyright 2019 Lorenzo"
-__version__ = "0.1.2a"
+__version__ = "0.1.3a"
 
 from collections import namedtuple
 
@@ -23,4 +23,4 @@ from .types import *  # noqa: F401
 
 VersionInfo = namedtuple("VersionInfo", "major minor micro releaselevel")
 
-version_info = VersionInfo(major=0, minor=1, micro=2, releaselevel="alpha")
+version_info = VersionInfo(major=0, minor=1, micro=3, releaselevel="alpha")