+++ async_pokepy/http.py
@@ -130,5 +130,8 @@ class HTTPPokemonClient:
     def get_ability(self, query: Union[int, str]) -> Coroutine:
         return self.request(Route(self.base, "ability", query))
 
+    def get_berry(self, query: Union[int, str]) -> Coroutine:
+        return self.request(Route(self.base, "berry", query))
+
     def get_pagination(self, query: str, **kwargs) -> Coroutine:
-        return self.request(Route(self.base, _fmt_param(query), **kwargs))
+        return self.request(Route(self.base, query, **kwargs))