+++ async_pokepy/types/__init__.py
@@ -26,6 +26,7 @@ DEALINGS IN THE SOFTWARE.
 
 from .abc import *  # noqa: F401
 from .ability import *  # noqa: F401
+from .berry import *  # noqa: F401
 from .common import *  # noqa: F401
 from .move import *  # noqa: F401
 from .pagination import *  # noqa: F401