+++ async_pokepy/types/common.py
@@ -31,7 +31,53 @@ __all__ = (
     "Effect",
     "VerboseEffect",
     "VersionGameIndex",
+    "APIObject",
+    "NamedAPIObject",
+    "MachineVersionDetail"
 )
+
+
+class APIObject:
+    """Represents a partial API object with an ID.
+
+    Attributes
+    ----------
+    id: :class:`int`
+        The object's identifier."""
+    def __init__(self, data: dict):
+        self.id = int(data["url"].split("/")[-2])
+
+    def __repr__(self) -> str:
+        return "<APIObject id={0.id}>".format(self)
+
+
+class NamedAPIObject(APIObject):
+    """Represents a partial API object with a name and ID.
+
+    This inherits from :class:`APIObject`.
+
+    .. container:: operations
+
+        .. describe:: str(x)
+
+            Returns the object's name.
+
+    Attributes
+    ----------
+    id: :class:`int`
+        The object's identifier.
+    name: :class:`str`
+        The object's name."""
+    def __init__(self, data: dict):
+        super().__init__(data)
+
+        self.name = _pretty_format(data["name"])
+
+    def __str__(self) -> str:
+        return self.name
+
+    def __repr__(self) -> str:
+        return "<NamedAPIObject id={0.id} name='{0}'>".format(self)
 
 
 class Name:
@@ -49,13 +95,13 @@ class Name:
     ----------
     name: :class:`str`
         The name.
-    language: :class:`str`
+    language: :class:`NamedAPIObject`
         The language in which the name is in."""
     __slots__ = ("name", "language")
 
     def __init__(self, data: dict):
         self.name = _pretty_format(data["name"])
-        self.language = data["language"]["name"]
+        self.language = NamedAPIObject(data["language"])
 
     def __str__(self) -> str:
         return self.name
@@ -75,14 +121,14 @@ class VerboseEffect:
         The localized effect text in the associated language.
     short_effect: :class:`str`
         The localized effect text in brief.
-    language: :class:`str`
+    language: :class:`NamedAPIObject`
         The language the effect is in."""
     __slots__ = ("effect", "short_effect", "language")
 
     def __init__(self, data: dict):
         self.effect = data["effect"]
         self.short_effect = data["short_effect"]
-        self.language = data["language"]["name"]
+        self.language = NamedAPIObject(data["language"])
 
     def __repr__(self) -> str:
         return "<VerboseEffect language='{0.language}'>".format(self)
@@ -94,14 +140,14 @@ class VersionGameIndex:
     Attributes
     ----------
     game_index: :class:`int`
-        The internal id of a PokeAPI resource within game data.
-    version: :class:`str`
+        The internal id of a PokeAPI object within game data.
+    version: :class:`NamedAPIObject`
         The name of the version relevant to the game index."""
     __slots__ = ("game_index", "version")
 
     def __init__(self, data: dict):
         self.game_index = data["game_index"]
-        self.version = _pretty_format(data["version"]["name"])
+        self.version = NamedAPIObject(data["version"])
 
     def __repr__(self) -> str:
         return "<VersionGameIndex game_index={0.game_index} version='{0.version}'>".format(self)
@@ -120,16 +166,33 @@ class Effect:
     ----------
     effect: :class:`str`
         The localized effect text in the associated language.
-    language: :class:`str`
+    language: :class:`NamedAPIObject`
         The language the effect is in."""
     __slots__ = ("effect", "language")
 
     def __init__(self, data: dict):
         self.effect = data["effect"]
-        self.language = data["language"]["name"]
+        self.language = NamedAPIObject(data["language"])
 
     def __str__(self) -> str:
         return self.effect
 
     def __repr__(self) -> str:
         return "<Effect language={0.language}>".format(self)
+
+
+class MachineVersionDetail:
+    """Represents the version details of a machine.
+
+    Attributes
+    ----------
+    machine: :class:`APIObject`
+        The machine that teaches a move from an item.
+    version_group: :class:`NamedAPIObject`
+        The version group of this specific machine."""
+    def __init__(self, data: dict):
+        self.machine = APIObject(data)
+        self.version_group = NamedAPIObject(data)
+
+    def __repr__(self) -> str:
+        return "<MachineVersionDetail version_group='{0.version_group}'>".format(self)